Moved Siesta iterators in separate file. Changed iterate_over management.
In the new file future_iterate I moved all the classes that will be, in my
vision, in the iterate file of the parent folder.
Also changed again the intenal organization of the functions parsing
the iterate_over input.
